telegram_bot/requirements.txt
# Telegram Bot
python-telegram-bot>=21.0,<22.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
cryptography>=42.0.0

# Existing dependencies (from main app)
telethon>=1.42.0
anthropic>=0.76.0
pandas>=2.3.0
python-docx>=1.2.0
python-dotenv>=1.2.0

# CORS
python-multipart>=0.0.9

# Async support
aiofiles>=23.2.1
